Exit randomisation and Inhibitors
