+++ UnityProject/Assets/Prefabs/Items/Research/AdminHandTele.prefab
@@ -14,6 +14,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 4933656907707707143, guid: e33403c0371fc174aa03640734665a15,
         type: 3}
+      propertyPath: ignoreInhibitors
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4933656907707707143, guid: e33403c0371fc174aa03640734665a15,
+        type: 3}
       propertyPath: trackingBeaconType
       value: 3
       objectReference: {fileID: 0}
