+++ UnityProject/Assets/Prefabs/Objects/Doors/MachineDoors/Airlocks/AirlockSandstoneOpaqueSelfPowered.prefab
@@ -1,0 +1,177 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &6772816149642416862
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 495997637782654300, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 532313921234287200, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: isWindowedDoor
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 1343151189207969991, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 1782512652456972366, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 1937390978069796087, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 2126275048737528191, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 2179906470425791467, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 2234499430741308333, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 2284870068917293999, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: _assetId
+      value: 2757735653
+      objectReference: {fileID: 0}
+    - target: {fileID: 2284870068917293999, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_AssetId
+      value: 4e421f8fcc4ec2e4f8d1c9c4bebe1387
+      objectReference: {fileID: 0}
+    - target: {fileID: 4022813679251449507, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 6145748976052171221, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Sprite
+      value: 
+      objectReference: {fileID: 21300000, guid: 6d50ca0dd8db2444cb58488d2e6b1b29,
+        type: 3}
+    - target: {fileID: 6573607110591482081, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: isSelfPowered
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6708468814386862212, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: foreverID
+      value: fd570f4a560f801408a9221fa8443ec6
+      objectReference: {fileID: 0}
+    - target: {fileID: 7667568523356956365, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Name
+      value: AirlockSandstoneOpaqueSelfPowered
+      objectReference: {fileID: 0}
+    - target: {fileID: 7667568523356956365, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_Layer
+      value: 17
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7673317071903629335, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8744020800374120540, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: PresentSpriteSet
+      value: 
+      objectReference: {fileID: 11400000, guid: 7518e5fe61096f54dbd2ebf4fb96e90f,
+        type: 2}
+    - target: {fileID: 8744020800374120540, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: SubCatalogue.Array.data[0]
+      value: 
+      objectReference: {fileID: 11400000, guid: 7518e5fe61096f54dbd2ebf4fb96e90f,
+        type: 2}
+    - target: {fileID: 8744020800374120540, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: SubCatalogue.Array.data[1]
+      value: 
+      objectReference: {fileID: 11400000, guid: 8b83e16372678f74b8db696e724450d9,
+        type: 2}
+    - target: {fileID: 8744020800374120540, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: SubCatalogue.Array.data[2]
+      value: 
+      objectReference: {fileID: 11400000, guid: 388a28aad0b8ff147a90b88812cf307b,
+        type: 2}
+    - target: {fileID: 8744020800374120540, guid: 86e0f7d4c8f79894aa974cec417094df,
+        type: 3}
+      propertyPath: SubCatalogue.Array.data[3]
+      value: 
+      objectReference: {fileID: 11400000, guid: 543f94346fb050b438f35674ab8ed857,
+        type: 2}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 86e0f7d4c8f79894aa974cec417094df, type: 3}
