+++ UnityProject/Assets/Prefabs/Objects/Doors/MachineDoors/Airlocks/AirlockSandstoneOpaqueSelfPowered.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fd570f4a560f801408a9221fa8443ec6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
