+++ UnityProject/Assets/Prefabs/SceneConstruction/MaintGenerator.prefab
@@ -57,20 +57,20 @@ MonoBehaviour:
     <RequireOpposingWalls>k__BackingField: 0
     <SpawnLockerCrate>k__BackingField: 0
     <TileToSpawn>k__BackingField: {fileID: 0}
-  - <ObjectToSpawn>k__BackingField: {fileID: 7376449702682419346, guid: e7a19fe59d5bcfe49a1d885d7fedab6e,
-      type: 3}
+  - <ObjectToSpawn>k__BackingField: {fileID: 0}
     <ObjectChance>k__BackingField: 10
     <RequiredWalls>k__BackingField: 2
     <RequireOpposingWalls>k__BackingField: 1
     <SpawnLockerCrate>k__BackingField: 0
-    <TileToSpawn>k__BackingField: {fileID: 0}
-  - <ObjectToSpawn>k__BackingField: {fileID: 5284686503705020292, guid: 0ce26a7341e66b142b36e308022db584,
-      type: 3}
+    <TileToSpawn>k__BackingField: {fileID: 11400000, guid: 3975642db6c52ed4fa2086247abc1f28,
+      type: 2}
+  - <ObjectToSpawn>k__BackingField: {fileID: 0}
     <ObjectChance>k__BackingField: 10
     <RequiredWalls>k__BackingField: 2
     <RequireOpposingWalls>k__BackingField: 0
     <SpawnLockerCrate>k__BackingField: 0
-    <TileToSpawn>k__BackingField: {fileID: 0}
+    <TileToSpawn>k__BackingField: {fileID: 11400000, guid: c87a62cb7a026433f9d4e361d2740cbc,
+      type: 2}
   - <ObjectToSpawn>k__BackingField: {fileID: 3378819989493369114, guid: c10109559e87d924bb10ee31d3bd33cb,
       type: 3}
     <ObjectChance>k__BackingField: 15
