--- UnityProject/Assets/Scripts/Systems/MaintGeneratorManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 474fd7e19918e2244a4aaf825e2511ad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
