+++ UnityProject/Assets/Scripts/Systems/MaintRooms/ExitMarker.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Systems.Scenes
+{
+	public class ExitMarker : MonoBehaviour, IServerSpawn
+	{
+		public void OnSpawnServer(SpawnInfo spawnInfo)
+		{
+			MaintGeneratorManager.possibleExits.Add(gameObject);
+		}
+	}
+}
