+++ UnityProject/Assets/Scripts/Systems/MaintRooms/ExitMarker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 82dad61adbf9f6142a11f5b7a3c00b76
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
