+++ UnityProject/Assets/Scripts/Systems/MaintRooms/MaintRoomsMarker.cs
@@ -1,0 +1,10 @@
+using Gateway;
+using UnityEngine;
+
+public class MaintRoomsMarker : MonoBehaviour, IServerSpawn
+{
+	public void OnSpawnServer(SpawnInfo spawnInfo)
+	{
+		TransportUtility.MaintRoomLocations.Add(this.gameObject);
+	}
+}
