+++ UnityProject/Assets/Scripts/Systems/MaintRooms/MaintRoomsMarker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7392328e383ac746a212578660bdf01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
