+++ UnityProject/Assets/Scripts/Systems/MaintRooms/TeleportInhibitor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 600d85cb1dca7e142903877e879415ed
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
