--- UnityProject/Assets/Scripts/Systems/MaintRoomsMarker.cs
@@ -1,10 +1,0 @@
-using Gateway;
-using UnityEngine;
-
-public class MaintRoomsMarker : MonoBehaviour, IServerSpawn
-{
-	public void OnSpawnServer(SpawnInfo spawnInfo)
-	{
-		TransportUtility.MaintRoomLocations.Add(this.gameObject);
-	}
-}
